Refactor delete method

+++ backend/src/main/java/com/github/raphaelfontoura/dscatalog/services/CategoryService.java
@@ -7,21 +7,22 @@ import com.github.raphaelfontoura.dscatalog.services.exceptions.DatabaseExceptio
 import com.github.raphaelfontoura.dscatalog.services.exceptions.ResourceNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.persistence.EntityNotFoundException;
-import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
-@RequiredArgsConstructor
 public class CategoryService {
 
     private final CategoryRepository repository;
+
+    public CategoryService(CategoryRepository repository) {
+        this.repository = repository;
+    }
 
     @Transactional(readOnly = true)
     public Page<CategoryDTO> findAll(Pageable pageable) {
@@ -56,11 +57,13 @@ public class CategoryService {
         }
     }
 
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw  new ResourceNotFoundException("Recurso nã encontrado");
+        }
         try {
             repository.deleteById(id);
-        } catch (EmptyResultDataAccessException e) {
-            throw new ResourceNotFoundException("Categoria não localizada. Id: " + id);
         } catch (DataIntegrityViolationException e) {
             throw new DatabaseException("Erro ao deletar registro do banco");
         }
